Refactor CbzPositionListFactory into PerResourcePositionListFactory, and use it for DiViNa
