+++ r2-streamer/src/main/java/org/readium/r2/streamer/parser/divina/DiViNaParser.kt
@@ -13,6 +13,7 @@ import org.json.JSONObject
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.streamer.BuildConfig.DEBUG
 import org.readium.r2.streamer.container.ContainerError
+import org.readium.r2.streamer.parser.PerResourcePositionListFactory
 import org.readium.r2.streamer.parser.PubBox
 import org.readium.r2.streamer.parser.PublicationParser
 import timber.log.Timber
@@ -86,10 +87,19 @@ class DiViNaParser : PublicationParser {
         val json = JSONObject(stringManifest)
 
         //Parsing manifest.json & building publication object
-        val publication = Publication.fromJSON(json)
-        publication?.type = Publication.TYPE.DiViNa
+        var publication = Publication.fromJSON(json)
+            ?: return null
 
-        return publication?.let {  PubBox(it, container) }
+        publication = publication.copy(
+            positionListFactory = PerResourcePositionListFactory(
+                readingOrder = publication.readingOrder,
+                fallbackMediaType = "image/*"
+            )
+        )
+
+        publication.type = Publication.TYPE.DiViNa
+
+        return PubBox(publication, container)
     }
 
     private fun InputStream.toFile(path: String) {
